% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{compare_qq}
\alias{compare_qq}
\title{QQ-plot of given two samples using ggplot2}
\usage{
compare_qq(sample1, sample2, main_title)
}
\arguments{
\item{sample1}{matrix of observations in sample 1}

\item{sample2}{matrix of observations in sample 2}

\item{main_title}{title of the generated plot}
}
\value{
QQ-plot of given samples
}
\description{
QQ-plot of given two samples using ggplot2
}
\keyword{internal}
