% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{compute_stats}
\alias{compute_stats}
\title{Compute some descriptive statistics for the two sample tests}
\usage{
compute_stats(var1, var2, eps = 3)
}
\arguments{
\item{var1}{vector of observations of a given variable from sample 1}

\item{var2}{vector of observations of a given variable from sample 2}

\item{eps}{precision of displayed statistics}
}
\value{
Computed statistics
}
\description{
Compute some descriptive statistics for the two sample tests
}
\keyword{internal}
