% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computePoissonMatrix}
\alias{computePoissonMatrix}
\title{Compute the Poisson kernel matrix between observations in a sample.}
\usage{
computePoissonMatrix(x_mat, rho)
}
\arguments{
\item{x_mat}{A matrix containing the observations.}

\item{rho}{concentration parameter of the Poisson kernel.}
}
\value{
Matrix of computed Poisson kernel values
}
\description{
Compute the Poisson kernel matrix between observations in a sample.
}
\examples{
x <- matrix(rnorm(300),ncol=3)
# Normalize each observation
x <- x/sqrt(rowSums(x^2))
computePoissonMatrix(x_mat = x, rho=0.8)

}
