% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{DOF}
\alias{DOF}
\title{Degrees of freedom (DOF) for the Poisson kernel}
\usage{
DOF(d, rho)
}
\arguments{
\item{d}{the number of dimensions}

\item{rho}{concentration parameter}
}
\value{
a list containing the DOF and the coefficient c of the asymptotic
distribution
}
\description{
Compute the Degrees of Freedom (DOF) of the Poisson Kernel given the
dimension d and concentration parameter rho
}
\keyword{internal}
