% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{extract_stats}
\alias{extract_stats}
\title{Descriptive statistics for the clusters identified by the Poisson
kernel-based clustering.}
\usage{
extract_stats(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{possible additional inputs}
}
\description{
Method for objects of class \code{pkbc} which computes some
descriptive for each variable with respect to the detected groups.
}
