% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kbNormTest}
\alias{kbNormTest}
\title{Compute kernel-based quadratic distance test for Normality}
\usage{
kbNormTest(x_mat, h, mu_hat, Sigma_hat, centeringType = "Param")
}
\arguments{
\item{x_mat}{A matrix containing the observations.}

\item{h}{The bandwidth parameter for the Gaussian kernel function.}

\item{mu_hat}{Mean vector for the reference distribution (if available)}

\item{Sigma_hat}{Covariance matrix of the reference distribution (if available)}

\item{centeringType}{String indicating the method used for centering the normal kernel ('Param' or 'Nonparam').}
}
\value{
A scalar value representing the test statistic.
}
\description{
Compute kernel-based quadratic distance test for Normality
}
\examples{
#x <- matrix(rnorm(100), ncol = 2)
#h=1

#mu_hat = matrix(rep(0,2),nrow=1)
#sigma_hat = diag(2)
#kbNormTest(x, h, mu_hat=mu_hat, Sigma_hat=sigma_hat ,centeringType = "Param")

}
\keyword{internal}
