% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NonparamCentering}
\alias{NonparamCentering}
\title{Non-parametric centered kernel}
\usage{
NonparamCentering(kmat_zz, n_z)
}
\arguments{
\item{kmat_zz}{A matrix containing the values of the kernel functions for each pair of observations}

\item{n_z}{The number of total observations}
}
\value{
Matrix of centered kernel
}
\description{
Compute the values of the non-parametric centered kernel given the values of the 
kernel matrix.
}
\examples{
h <- 0.5
H <- h^2*diag(2)
x <- matrix(rnorm(100), ncol = 2) #50 x 2
y <- matrix(rnorm(100), ncol = 2) #50 x 2
k_mat <- computeKernelMatrix(x, y, H) #50 x 50
NonparamCentering(k_mat, 50)


}
