% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ParamCentering}
\alias{ParamCentering}
\title{Parametric centered kernel}
\usage{
ParamCentering(kmat_zz, z_mat, H, mu_hat, Sigma_hat)
}
\arguments{
\item{kmat_zz}{A matrix containing the values of the kernel functions for each pair of observations}

\item{z_mat}{Matrix of observations used for computing the kernel matrix}

\item{H}{Covariance matrix of the Normal kernel}

\item{mu_hat}{Mean of centering of Normal distribution}

\item{Sigma_hat}{Covariance of centering Normal distribution}
}
\value{
Matrix of centered kernel
}
\description{
Compute the Gaussian kernel centered with respect to a Normal distribution
with mean vector mu and covariance Sigma
}
\examples{
#h <- 0.5
#H <- h^2*diag(2)
#x <- matrix(rnorm(100), ncol = 2)
#y <- matrix(rnorm(100), ncol = 2)
#z <- rbind(x, y)
#mu_hat <- matrix(colMeans(z),nrow=1)
#Sigma_hat <- cov(z)
#k_mat <- computeKernelMatrix(z, z, H)
#ParamCentering(k_mat, z, H, mu_hat, Sigma_hat)


}
\keyword{internal}
