% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{pkbc-class}
\alias{pkbc-class}
\title{A S4 class for the clustering algorithm on the sphere based on
Poisson kernel-based distributions.}
\description{
A class to represent the results of Poisson kernel-based
clustering procedure for spherical observations.
}
\details{
See the function \code{pkbc} for more details.
}
\section{Slots}{

\describe{
\item{\code{res_k}}{List of objects with the results of the clustering algorithm for
each value of possible number of clusters considered.}

\item{\code{input}}{List of input data}
}}

\examples{
data("wireless")
res <- pkbc(as.matrix(wireless[,-8]),4)

}
