% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sample_hypersphere}
\alias{sample_hypersphere}
\title{Generate random Sample from the hypersphere}
\usage{
sample_hypersphere(d, n_points = 1)
}
\arguments{
\item{d}{number of dimensions.}

\item{n_points}{number of sampled observations.}
}
\value{
data matrix with the sampled observations.
}
\description{
Generate random Sample from the hypersphere
}
\examples{
x_sp <- sample_hypersphere(3,100)

}
