% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stat2sample}
\alias{stat2sample}
\title{Compute kernel-based quadratic distance two-sample test with Normal kernel}
\usage{
stat2sample(x_mat, y_mat, h, mu_hat, Sigma_hat, centeringType = "Nonparam")
}
\arguments{
\item{x_mat}{A matrix containing observations from the first sample}

\item{y_mat}{A matrix containing observations from the second sample}

\item{h}{The bandwidth parameter for the kernel function}

\item{mu_hat}{Mean vector for the reference distribution}

\item{Sigma_hat}{Covariance matrix of the reference distribution}

\item{centeringType}{String indicating the method used for centering the normal kernel}
}
\value{
A scalar value representing the test statistic
}
\description{
Compute kernel-based quadratic distance two-sample test with Normal kernel
}
\details{
\code{mu_hat} and \code{Sigma_hat} need to be provided even if they are used 
only in case of "Param" \code{centeringType}.
}
\examples{
x <- matrix(rnorm(100), ncol = 2)
y <- matrix(rnorm(100), ncol = 2)
mu_hat <- matrix(0,nrow=1)
sigma_hat <- diag(2)
stat2sample(x, y, h = 1, mu_hat=mu_hat, Sigma_hat=sigma_hat)

}
