% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stat_ksample_cpp}
\alias{stat_ksample_cpp}
\title{Kernel-based quadratic distance k-sample tests}
\usage{
stat_ksample_cpp(x, y, h, sizes, cum_size)
}
\arguments{
\item{x}{A matrix containing observations from the pooled sample, from the k samples.}

\item{y}{A vector containing observations' memberships to the k samples.}

\item{h}{The bandwidth parameter for the kernel function.}

\item{sizes}{Vector with sample sizes of the considered samples}

\item{cum_size}{Vector indicating the cumulative sizes, adding one sample at the time.}
}
\value{
A vector containing the two k-sample test statistics
}
\description{
Compute the kernel-based quadratic distance k-sample tests with the Normal kernel and bandwidth parameter h.
}
\examples{
x1 <- matrix(rnorm(200), ncol = 2)
x2 <- matrix(rnorm(200), ncol = 2)
x3 <- matrix(rnorm(200, 2), ncol = 2)
x <- rbind(x1, x2, x3)
y <- rep(c(1,2,3), each=100)
sizes <- as.vector(table(y))
cum_size <- c(0,cumsum(sizes))
stat_ksample_cpp(x, y, h = 1, sizes, cum_size)

}
