% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{statPoissonUnif}
\alias{statPoissonUnif}
\title{Poisson kernel-based test for Uniformity on the Sphere}
\usage{
statPoissonUnif(x_mat, rho)
}
\arguments{
\item{x_mat}{A matrix containing the observations.}

\item{rho}{Concentration parameter of the Poisson kernel.}
}
\value{
Vector with the values of the U-statistic and V-statistic
}
\description{
Compute the Poisson kernel-based test for Uniformity given a sample of observations on the Sphere
}
\examples{
#x <- matrix(rnorm(300),ncol=3)
# Normalize each observation
#x <- x/sqrt(rowSums(x^2))
#statPoissonUnif(x, rho=0.7)

}
\keyword{internal}
