% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{summary,pkbc-method}
\alias{summary,pkbc-method}
\title{Summarizing PKBD mixture Fits}
\usage{
\S4method{summary}{pkbc}(object)
}
\arguments{
\item{object}{Object of class \code{pkbc}}
}
\description{
Summary method for class "pkbc"
}
\examples{
dat <- rbind(matrix(rnorm(100),2),matrix(rnorm(100,5),2))
res <- pkbc(dat,2:4)
summary(res)

}
\seealso{
\code{\link[=pkbc]{pkbc()}}
}
