% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{var_k}
\alias{var_k}
\title{Exact variance of k-sample test}
\usage{
var_k(Kcen, sizes, cum_size)
}
\arguments{
\item{Kcen}{the matrix with centered kernel values}

\item{sizes}{vector indicating sample's size.}

\item{cum_size}{vector indicating sample's cumulative sizes.}
}
\value{
the value of computed variance.
}
\description{
Compute the exact variance of kernel test for the k-sample problem under
the null hypothesis that F1=...=Fk.
}
\keyword{internal}
