% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{var_norm}
\alias{var_norm}
\title{Exact variance of normality test}
\usage{
var_norm(d, h, n)
}
\arguments{
\item{d}{the number of dimensions}

\item{h}{bandwidth parameter}

\item{n}{sample size}
}
\value{
the value of computed variance
}
\description{
Compute the exact variance of kernel test for normality under the null
hypothesis that G=N(0,I).
}
\keyword{internal}
