% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{var_two}
\alias{var_two}
\title{Exact variance of two-sample test}
\usage{
var_two(Kcen, nsamples)
}
\arguments{
\item{Kcen}{the matrix with centered kernel values}

\item{nsamples}{vector indicating sample's membership.}
}
\value{
the value of computed variance.
}
\description{
Compute the exact variance of kernel test for the two-sample problem under
the null hypothesis that F=G.
}
\keyword{internal}
