% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{compute_stats}
\alias{compute_stats}
\title{Compute and display some descriptive statistics for the two sample tests}
\usage{
compute_stats(var1, var2, var_name, eps = 3)
}
\arguments{
\item{var1}{vector of observations of a given variable from sample 1}

\item{var2}{vector of observations of a given variable from sample 2}

\item{var_name}{Name of the variable displayed}

\item{eps}{precision of displayed statistics}
}
\value{
Computed statistics with a plot
}
\description{
Compute and display some descriptive statistics for the two sample tests
}
\keyword{internal}
