% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeKernelMatrix}
\alias{computeKernelMatrix}
\title{Compute the Gaussian kernel matrix between two samples}
\usage{
computeKernelMatrix(x_mat, y_mat, H)
}
\arguments{
\item{x_mat}{A matrix containing the observations of X}

\item{y_mat}{A matrix containing the observations of Y}

\item{H}{Covariance matrix of the Gaussian kernel}
}
\value{
Matrix of computed Gaussian kernel values
}
\description{
Compute the Gaussian kernel matrix between two samples
}
\examples{
h <- 0.5
H <- h^2*diag(2)
x <- matrix(rnorm(100), ncol = 2)
y <- matrix(rnorm(100), ncol = 2)
computeKernelMatrix(x, y, H)

}
