% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{generate_SN}
\alias{generate_SN}
\title{Generate two samples data from skew-normal distributions}
\usage{
generate_SN(d, size_x, size_y, mu_x, mu_y, sigma_x, sigma_y, skewness_y)
}
\arguments{
\item{d}{number of dimensions.}

\item{size_x}{the number of observations for sample X}

\item{size_y}{the number of observations for sample Y}

\item{mu_x}{the mean of X}

\item{mu_y}{the mean of Y}

\item{sigma_x}{the standard deviation of X}

\item{sigma_y}{the standard deviation of Y}

\item{skewness_y}{the skewness of Y (the skewness of X is set to zero).}
}
\value{
a list containing the generated X and Y data sets.
}
\description{
This function generates data from skew-normal distributions with the
specified parameters of means and covariance matrices.
}
\keyword{internal}
