% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{var_norm}
\alias{var_norm}
\title{Exact variance of normality test}
\usage{
var_norm(Sigma_h, V, n)
}
\arguments{
\item{Sigma_h}{covariance matrix of the gaussian kernel}

\item{V}{Covariance matrix of the tested distribution G}

\item{n}{sample size}
}
\value{
the value of computed variance
}
\description{
Compute the exact variance of kernel test for normality under the null
hypothesis that G=N(0,I).
}
\keyword{internal}
